requirements.txt
# Core dependencies
python-dotenv==1.0.0
pandas==2.1.0

# Testing dependencies
pytest==7.4.0
pytest-cov==4.1.0

# Code quality
pylint==2.17.5
